chore: move critical wo table to app
