+++ src/apps/Construction/Components/CriticalWoTable/components/WoNumbers.tsx
@@ -2,8 +2,8 @@ import styled from 'styled-components';
 import { WoMapCount } from '..';
 import { tokens } from '@equinor/eds-tokens';
 import { useCallback } from 'react';
-import { openSidesheet } from '../../../../../Sidesheet/Functions';
-import { SidesheetContent } from '../../../Components';
+import { SidesheetContent } from '@equinor/Diagrams';
+import { openSidesheet } from '@equinor/sidesheet';
 
 enum CircleColor {
     WARNING,
@@ -47,12 +47,9 @@ export const WoNumbersDisplay = <T extends unknown>({
     groupedKey,
     keysOfFiltered,
 }: WoNumbersDisplayProps<T>) => {
-    const onClick = useCallback(
-        (item: T[]) => {
-            openSidesheet(SidesheetContent, { data: item });
-        },
-        [openSidesheet]
-    );
+    const onClick = useCallback((item: T[]) => {
+        openSidesheet(SidesheetContent, { data: item });
+    }, []);
     return (
         <>
             {Object.values(filtered[groupedKey]).map((item, index) => {
