+++ src/apps/Construction/Components/DetailsPage/utils.tsx
@@ -1,4 +1,4 @@
-import { CellProps, TableInstance } from '@equinor/Table';
+import { CellProps, TableInstance, Column } from '@equinor/Table';
 import {
     SumColumnFooter,
     SumColumnFooterCount,
@@ -15,7 +15,7 @@ import { ColumnGeneratorArgs } from './types';
  */
 export const columnGenerator = <T extends Record<string, unknown>>(
     args: ColumnGeneratorArgs<T>
-) => {
+): Column<T> => {
     const { id, header, accessorKey, aggregate, footerType, jobStatus, width = 100 } = args;
     return {
         id,
